[FIX] style errors
